Move mock to a new home
